+++ WorkoutManager.BusinessLogic/DTOs/SessionDtos.cs
@@ -2,8 +2,8 @@ namespace WorkoutManager.BusinessLogic.DTOs;
 
 public class SessionSummaryDto
 {
-    public int Id { get; set; }
-    public int PlanId { get; set; }
+    public long Id { get; set; }
+    public long PlanId { get; set; }
     public string? PlanName { get; set; }
     public string? TrainingDayName { get; set; }
     public string? Notes { get; set; }
@@ -13,7 +13,7 @@ public class SessionSummaryDto
 
 public class SessionDetailsDto
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public string? Notes { get; set; }
     public DateTime StartTime { get; set; }
     public DateTime? EndTime { get; set; }
@@ -22,8 +22,8 @@ public class SessionDetailsDto
 
 public class SessionExerciseDetailsDto
 {
-    public int Id { get; set; }
-    public int ExerciseId { get; set; }
+    public long Id { get; set; }
+    public long ExerciseId { get; set; }
     public string? Notes { get; set; }
     public bool Skipped { get; set; }
     public int Order { get; set; }
@@ -47,7 +47,7 @@ public class UpdateSessionExerciseDto
 
 public class ExerciseSetDto
 {
-    public int Id { get; set; }
+    public long Id { get; set; }
     public decimal Weight { get; set; }
     public int Reps { get; set; }
     public bool IsFailure { get; set; }
