+++ WorkoutManager.BusinessLogic/Services/Interfaces/IExerciseRepository.cs
@@ -6,8 +6,8 @@ namespace WorkoutManager.BusinessLogic.Services.Interfaces;
 public interface IExerciseRepository
 {
     Task<IEnumerable<Exercise>> GetExercisesForUserAsync(Guid userId);
-    Task<Exercise?> GetExerciseByIdAsync(int exerciseId);
+    Task<Exercise?> GetExerciseByIdAsync(long exerciseId);
     Task<Exercise?> GetExerciseByNameForUserAsync(string name, Guid userId);
     Task<Exercise> CreateExerciseAsync(Exercise exercise);
-    Task<PreviousExercisePerformanceDto?> GetLastPerformanceAsync(int exerciseId, Guid userId);
+    Task<PreviousExercisePerformanceDto?> GetLastPerformanceAsync(long exerciseId, Guid userId);
 }
