+++ WorkoutManager.BusinessLogic/Services/Interfaces/IExerciseService.cs
@@ -7,14 +7,14 @@ public interface IExerciseService
     Task<PaginatedList<ExerciseDto>> GetExercisesAsync(
         Guid userId,
         string? search = null,
-        int? muscleGroupId = null,
+        long? muscleGroupId = null,
         int page = 1,
         int pageSize = 20);
-    
-    Task<ExerciseDto?> GetExerciseByIdAsync(int exerciseId);
-    
+
+    Task<ExerciseDto?> GetExerciseByIdAsync(long exerciseId);
+
     Task<ExerciseDto> CreateExerciseAsync(CreateExerciseDto dto, Guid userId);
-    
-    Task<PreviousExercisePerformanceDto?> GetLastPerformanceAsync(int exerciseId, Guid userId);
+
+    Task<PreviousExercisePerformanceDto?> GetLastPerformanceAsync(long exerciseId, Guid userId);
 }
