+++ WorkoutManager.BusinessLogic/Services/Interfaces/IMuscleGroupService.cs
@@ -5,6 +5,6 @@ namespace WorkoutManager.BusinessLogic.Services.Interfaces;
 public interface IMuscleGroupService
 {
     Task<IEnumerable<MuscleGroupDto>> GetAllMuscleGroupsAsync();
-    Task<MuscleGroupDto?> GetMuscleGroupByIdAsync(int id);
+    Task<MuscleGroupDto?> GetMuscleGroupByIdAsync(long id);
 }
