+++ WorkoutManager.BusinessLogic/Services/Interfaces/ISessionExerciseRepository.cs
@@ -7,12 +7,12 @@ namespace WorkoutManager.BusinessLogic.Services.Interfaces;
 
 public interface ISessionExerciseRepository
 {
-    Task<Session?> GetSessionByIdAndUserIdAsync(int sessionId, Guid userId);
-    Task<SessionExercise?> GetSessionExerciseByIdAndSessionIdAsync(int sessionExerciseId, int sessionId);
-    Task<SessionExercise?> GetSessionExerciseWithSessionAsync(int sessionExerciseId, Guid userId);
+    Task<Session?> GetSessionByIdAndUserIdAsync(long sessionId, Guid userId);
+    Task<SessionExercise?> GetSessionExerciseByIdAndSessionIdAsync(long sessionExerciseId, long sessionId);
+    Task<SessionExercise?> GetSessionExerciseWithSessionAsync(long sessionExerciseId, Guid userId);
     Task UpdateSessionExerciseAsync(SessionExercise sessionExercise);
-    Task DeleteSetsForSessionExerciseAsync(int sessionExerciseId);
-    Task<IEnumerable<ExerciseSet>> AddSetsToSessionExerciseAsync(int sessionExerciseId, IEnumerable<ExerciseSet> sets);
+    Task DeleteSetsForSessionExerciseAsync(long sessionExerciseId);
+    Task<IEnumerable<ExerciseSet>> AddSetsToSessionExerciseAsync(long sessionExerciseId, IEnumerable<ExerciseSet> sets);
 }
