+++ WorkoutManager.BusinessLogic/Services/Interfaces/ISessionRepository.cs
@@ -9,14 +9,14 @@ namespace WorkoutManager.BusinessLogic.Services.Interfaces;
 public interface ISessionRepository
 {
     Task<bool> HasActiveSessionAsync(Guid userId);
-    Task<TrainingDay?> GetTrainingDayByIdAsync(int trainingDayId);
+    Task<TrainingDay?> GetTrainingDayByIdAsync(long trainingDayId);
     Task<WorkoutPlan?> GetWorkoutPlanByIdAsync(long planId, Guid userId);
     Task<Session> CreateSessionAsync(Session session);
-    Task<IEnumerable<PlanDayExercise>> GetPlanDayExercisesAsync(int trainingDayId);
+    Task<IEnumerable<PlanDayExercise>> GetPlanDayExercisesAsync(long trainingDayId);
     Task<SessionExercise> CreateSessionExerciseAsync(SessionExercise sessionExercise);
     Task<IEnumerable<Session>> GetSessionHistoryAsync(Guid userId, int page, int pageSize);
-    Task<Session?> GetSessionByIdAsync(int sessionId, Guid userId);
-    Task<IEnumerable<SessionExercise>> GetSessionExercisesWithSetsAsync(int sessionId);
+    Task<Session?> GetSessionByIdAsync(long sessionId, Guid userId);
+    Task<IEnumerable<SessionExercise>> GetSessionExercisesWithSetsAsync(long sessionId);
     Task UpdateSessionAsync(Session session);
     Task<Session?> GetActiveSessionAsync(Guid userId);
 }
