+++ WorkoutManager.BusinessLogic/Services/Interfaces/ISessionService.cs
@@ -5,11 +5,11 @@ namespace WorkoutManager.BusinessLogic.Services.Interfaces;
 
 public interface ISessionService
 {
-    Task<SessionDetailsDto> StartSessionAsync(int trainingDayId, Guid userId);
+    Task<SessionDetailsDto> StartSessionAsync(long trainingDayId, Guid userId);
     Task<PaginatedList<SessionSummaryDto>> GetSessionHistoryAsync(Guid userId, int page = 1, int pageSize = 20);
-    Task<SessionDetailsDto> GetSessionByIdAsync(int sessionId, Guid userId);
-    Task UpdateSessionNotesAsync(int sessionId, string? notes, Guid userId);
-    Task FinishSessionAsync(int sessionId, string? notes, Guid userId);
+    Task<SessionDetailsDto> GetSessionByIdAsync(long sessionId, Guid userId);
+    Task UpdateSessionNotesAsync(long sessionId, string? notes, Guid userId);
+    Task FinishSessionAsync(long sessionId, string? notes, Guid userId);
     Task<bool> HasActiveSessionAsync(Guid userId);
     Task<SessionDetailsDto?> GetActiveSessionAsync(Guid userId);
 }
