+++ WorkoutManager.Web/Services/ISessionService.cs
@@ -6,11 +6,11 @@ namespace WorkoutManager.Web.Services
     public interface ISessionService
     {
         Task<PaginatedList<SessionSummaryDto>> GetSessionHistoryAsync();
-        Task<SessionDetailsDto> GetSessionDetailsAsync(int id);
-        Task<SessionDetailsDto> StartSessionAsync(int trainingDayId);
-        Task UpdateSessionExerciseAsync(int sessionId, int sessionExerciseId, UpdateSessionExerciseDto payload);
-        Task UpdateSessionAsync(int sessionId, UpdateSessionCommand command);
-        Task FinishSessionAsync(int sessionId, string? notes);
+        Task<SessionDetailsDto> GetSessionDetailsAsync(long id);
+        Task<SessionDetailsDto> StartSessionAsync(long trainingDayId);
+        Task UpdateSessionExerciseAsync(long sessionId, long sessionExerciseId, UpdateSessionExerciseDto payload);
+        Task UpdateSessionAsync(long sessionId, UpdateSessionCommand command);
+        Task FinishSessionAsync(long sessionId, string? notes);
         Task<SessionDetailsDto?> GetActiveSessionAsync();
     }
 }
